navigation-api/navigation-methods/return-value/back-204-205-download.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script src="/common/utils.js"></script>

<body>
<script>
const tests = [
  ["204s", "204"],
  ["205s", "205"],
  ["Content-Disposition: attachment responses", "download"]
];

for (const [description, action] of tests) {
  promise_test(async t => {
    const id = token();

    const i = document.createElement("iframe");
    i.src = `resources/204-205-download-on-second-visit.py?id=${id}`;
    document.body.append(i);
    await new Promise(r => i.onload = r);

    // Configure it to return a 204 on the next visit
    await fetch(i.src + `&action=${action}`, { method: "POST" });

    // Now navigate elsewhere
    i.contentWindow.location.href = "/common/blank.html";
    await new Promise(r => i.onload = r);

    // Now try going back. It should do nothing (and not tell us about the result).

    const indexBefore = i.contentWindow.navigation.currentEntry.index;

    let navigateCount = 0;
    i.contentWindow.navigation.onnavigate = () => { ++navigateCount; };
    i.contentWindow.navigation.onnavigatesuccess = t.unreached_func("onnavigatesuccess should not be called");
    i.contentWindow.navigation.onnavigateerror = t.unreached_func("onnavigateerror should not be called");

    const result = i.contentWindow.navigation.back();

    assert_equals(navigateCount, 0, "navigate must not fire synchronously for traversals");
    assertNeverSettles(t, result, i.contentWindow);

    await new Promise(resolve => t.step_timeout(resolve, 100));
    assert_equals(navigateCount, 1, "navigate must have fired eventually, even though nothing else did");
    assert_equals(i.contentWindow.navigation.currentEntry.index, indexBefore);
  }, `back() promises to ${description} never settle`);
}
</script>
